add images and load src by js
