+++ src/styles/newCarCarousel.css
@@ -1,0 +1,86 @@
+.micro-slider {
+  height: 480px;
+  position: relative;
+  width: 100%;
+}
+
+.micro-slider.fullwidth {
+  height: 480px;
+}
+.micro-slider.fullwidth .slider-item {
+  height: 480px;
+  line-height: 480px;
+  width: 100%;
+}
+.slider-wrapper {
+  overflow: hidden;
+  position: relative;
+  transform-style: preserve-3d;
+  transform-origin: 0 50%;
+}
+.slider-item {
+  background: #fff;
+  box-shadow:
+    0 1px 3px rgba(0, 0, 0, 0.06),
+    0 1px 2px rgba(0, 0, 0, 0.12);
+  color: #fff;
+  display: none;
+  font-size: 72px;
+  height: 320px;
+  line-height: 320px;
+  left: 0;
+  position: absolute;
+  text-align: center;
+  top: 0;
+  width: 320px;
+  cursor: pointer;
+}
+.indicators {
+  bottom: 0px;
+  left: 50%;
+  list-style-type: none;
+  margin: 0 0 0 -77px;
+  padding: 0;
+  position: absolute;
+}
+.indicators li {
+  color: #fff;
+  float: left;
+  height: 16px;
+  margin-right: 8px;
+  text-align: center;
+  width: 16px;
+}
+.indicators li:last-child {
+  margin: 0;
+}
+.indicators a {
+  background: #fff;
+  border-radius: 8px;
+  border: 1px solid #e6e9ec;
+  color: #fff;
+  display: inline-block;
+  height: 16px;
+  width: 16px;
+}
+.indicators .active a {
+  background: #e6e9ec;
+}
+.s1 {
+  background: #3f51b5;
+}
+.s2 {
+  background: #448aff;
+}
+.s3 {
+  background: #03a9f4;
+}
+.s4 {
+  background: #00bcd4;
+}
+.s5 {
+  background: #009688;
+}
+.s6 {
+  background: #4caf50;
+}
